Add exception handling when k8s can't delete a cronjob.

+++ apis/resources/kubernetesClient.py
@@ -2,6 +2,7 @@ import base64
 import json
 import numbers
 import os
+import time
 from datetime import datetime, timezone
 
 import yaml
@@ -426,21 +427,16 @@ class ApiK8sClient:
 
     # Cron Job
     def list_namespaced_cron_job(self, namespace):
-        try:
-            return self.batch_v1beta1.list_namespaced_cron_job(namespace)
-        except apiError.DevOpsError as e:
-            if e.status_code != 404:
-                raise e
+        return self.batch_v1beta1.list_namespaced_cron_job(namespace)
 
     def delete_namespaced_cron_job(self, name, namespace):
-        try:
-            return self.batch_v1beta1.delete_namespaced_cron_job(name, namespace)
-        except apiError.DevOpsError as e:
-            if e.status_code != 404:
-                raise e
+        return self.batch_v1beta1.delete_namespaced_cron_job(name, namespace)
 
     def get_api_client(self):
         return self.api_k8s_client
+
+
+MAX_RETRY_APPLY_CRONJOB = 30
 
 
 def apply_cronjob_yamls():
@@ -454,13 +450,20 @@ def apply_cronjob_yamls():
                         if cronjob_json.metadata.name == json_file["metadata"]["name"]:
                             api_k8s_client.delete_namespaced_cron_job(cronjob_json.metadata.name,
                                                                       "default")
-                            while True:
+                            retry = 0
+                            while retry < MAX_RETRY_APPLY_CRONJOB:
+                                time.sleep(1)
                                 still_has_cj = False
                                 for cj in api_k8s_client.list_namespaced_cron_job("default").items:
                                     if cronjob_json.metadata.name in cj.metadata.name:
                                         still_has_cj = True
                                 if still_has_cj is False:
                                     break
+                                retry += 1
+                            if retry >= MAX_RETRY_APPLY_CRONJOB:
+                                logger.critical(f'Cannot delete existing cronjob {cronjob_json.metadata.name}!'
+                                                f' Cronjob is not applied.')
+                                return
                             for j in api_k8s_client.list_namespaced_job("default").items:
                                 if f"{cronjob_json.metadata.name}-" in j.metadata.name:
                                     api_k8s_client.delete_namespaced_job(
@@ -473,7 +476,9 @@ def apply_cronjob_yamls():
                     k8s_utils.create_from_yaml(
                         api_k8s_client.get_api_client(), os.path.join(root, file))
                 except k8s_utils.FailToCreateError as e:
+                    print('e1')
                     info = json.loads(e.api_exceptions[0].body)
+                    print(f'e2 {info}')
                     if info.get('reason').lower() == 'alreadyexists':
                         pass
                     else:
